requirements.txt
annotated-types==0.7.0
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
bcrypt==4.1.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.2
fastapi==0.115.0
fastapi-users==13.0.0
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
huggingface-hub==0.25.1
idna==3.10
Jinja2==3.1.4
makefun==1.15.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.1
opencv-python-headless==4.10.0.84
packaging==24.1
pillow==10.4.0
pwdlib==0.2.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.8.0
PyMuPDF==1.24.10
PyMuPDFb==1.24.10
pytesseract==0.3.13
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rsa==4.9
safetensors==0.4.5
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
sympy==1.13.3
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
transformers==4.45.1
typing_extensions==4.12.2
urllib3==2.2.3
uuid==1.30
uvicorn==0.31.0
